3rd Commit
New interfaces and types have been added into app dataObjects folder: iFormField, standardInputType, dynElementType, iformFieldOptionalValue, iformFieldValidator. The  ItemsTableFormFields: iformField[] const has benn also created and added in the same folder.

The formDynamicComponents new app folder has been created, and the following components have been added there: FormButtonDynamicElementComponent, FormButtonDynamicElementComponent, FormContainerComponent,  DynFormElementDirective, DynComponentFormElementService, and the DynamicFormModule barrel module, which exports only the FormContainerComponent. The DynamicFormModule has been also added in imports section in the AppModule.

The new component ‘ItemFormHostComponent’ has been created.

The ItemFormHostComponent is the parent (super parent) component of the FormContainerComponent. It defines and prepares the itemsFormFieldsSet1, a subset of the ItemsTableFormFields, of the form control fields that will appear in the dynamic form, as well as it adds a submit button. It passes down the itemsFormFieldsSet1 to its child-component, the FormContainerComponent, via property binding syntax in its remplate.

The FormContainerComponent creates (via using the formBuilder) the formGroup array of controls, on the basis of the set of fields received from its parent (ItemFormHostComponent). In its remplate, it loops through the the set of the fields via ngFor inside an ng-container, and for each field item it uses/applies the dynFormElementDirective passing it the field item, as well as the fornGroup.

The DynFormElementDirective has defined a constant (key-value pairs) of the candidate dynamic form components (the key is the type and value of the component). Then, (for each iteration),  it checks if the field passed in has a type property that matches the a key and if this is true creates the dynamic component (via the viewContainerRef), instantiates it (it uses the componentRef) and passes to the instance the field item, and the fornGroup, that have been received fromits parent (FormContainerComponent). It finally uses the ChangeDetectorRef to detect changes.