5th Commit branch master
We passed down, from item-query component, the item values obtained from the backend, into the ItemFormHostComponent. and use them as initial values to the subset of fields. This is actually, an object based on the IItem interface. We used the item object via property binding in the item-query template.
We added a specific (new) property -the formElementInitialValue property- (in the iformField interface). We use this new property to set the initial values of the form controls. We do this inside the prepareItemsFormFields() method of the item-form-host component.
We use a method handler, the onFormSubmit() method, for the form submission.
We have to patch the formGroup control values, whenever a change to a control occurs, otherwise only initial values (if they have been set) are submitted, and all other changed values are not submitted – are lost. We do this by subscribing to valueChanges into the ngOnChanges() and patching the form value with the changes.