6th Commit branch master
A new candidate dynamic component has been created the formDynamicComponents/form-select-dynamic-element. It is also has been added to the dynComponents of the DynFormElementDirective
For now the field is added separetely in the  itemsFormFieldsSet1, as well we manually add some of the options for test purposes in the item-form-host component.

Code improvements in the form-container component to solve submitting issues.
I call the createFornGroup() only once from within the ngOnInit. We can also subscribe inside the ngOnInit (it’s better than subscribing repeatedly in the ngOnChanges) for value changes, but this not necessary any more.
The createFornGroup() just creates once the form controls.
To handle the @Input changes (i.e. for getting the initial values from the item fetched from the backend) we use the ngOnChanges lifecycle hook, where we call the initializeFormControls().
The  initializeFormControls() method actually, uses the setValue() to set the initial values of the form controls (if any) and marks the formGroup as dirty. This causes the submit to include all values.