7th Commit branch master
The new ICategory interface has been added it in the dataObjects package folder.
The getCategories() new method has been added to the ItemsDataService that creates an Observable for fetching the categories from the backend.
The fetchItemCategories() method has been added to the ItemsDataService. It calls the getCategories() method in the ItemsDataService, fetches all Categories and creates an array of iformFieldOptionalValue objects. It assigns it to the component property itemCategories.
This property is passed down to the children ItemFormHostComponent as its fetchedItemCategories property.
We use this property, inside the prepareItemsFormFields(), to define the option-values for the configuration object for the select component.