8th Commit branch master
I added the missing categoryNames property for item category names in the IItem interface
I added the boolean valuePreselected property in the iformFieldOptionalValue interface
For any match we set the valuePreselected to true.
I set initial, preselected value in the form-container component for both of the methods: createFornGroup() and initializeFormControls(), just for the “select” components (form controls), in a similar way I did it before for the “input” ones.