16th Commit branch master
I created new material themes 1 for the entire app and 1 other for a component. Here I use it with the dateTime picker component. For this, I created a new SASS file the them1.scss – it uses the newer syntax using namespaces with '@use' instead of the older '@import'. Then I added it to the styles section of the angular.json file. I also have created the theme1old.scss file which uses the older syntax with '@import' instead of the newer '@use' - without defining a namespace e.g. 'mat'. The theme1old has not been added to angular.json.

I changed the templates (and/or other components code) for Material components implementation:
I updated the item-query
I added the mat-card as a container for entire dynamic form
I updated the form-input-dynamic-element
I updated the form-button-dynamic-element
I updated the form-radiobutton-dynamic-element
I updated the form-select-dynamic-element

I installed the @matero/extensions and I used the MtxDatetimepickerNodule with the MtxMomentDatetimeModule for the FormDatetimeDynamicElementComponent (orm-datetime-dynamic-element). I updated the material.module.ts accordingly.